Add PeerAddrStr and LocalAddrStr fields to bfdSession.
Eliminates the need to use inet_ntoa() to convert to a string each time
we want to log session id information.

The reason for adding this was to allow reporting both peer and local
addresses in a single call to bfdLog(). The inet_ntoa() function returns
a pointer to a static buffer, so when called more than once in an
argument list you get the value from the last invocation in all places
it was passed.